+++ resources/views/components/search/result-table.blade.php
@@ -30,14 +30,16 @@
 
             @if ($rows->count() > config('search.cap'))
                 <x-slot:foot>
-                    <x-table-cell colspan="4" align="right" sm>
-                        <a
-                            href="{{ route('projects.search.domain', ['project' => $project, 'domain' => $domain->value, 'q' => $query, 'mode' => $mode->value]) }}"
-                            class="text-link underline hover:text-link-hover"
-                        >
-                            {{ __('See all :count results', ['count' => $rows->count()]) }}
-                        </a>
-                    </x-table-cell>
+                    <tr>
+                        <x-table-cell colspan="4" align="right" sm>
+                            <a
+                                href="{{ route('projects.search.domain', ['project' => $project, 'domain' => $domain->value, 'q' => $query, 'mode' => $mode->value]) }}"
+                                class="text-link underline hover:text-link-hover"
+                            >
+                                {{ __('See all :count results', ['count' => $rows->count()]) }}
+                            </a>
+                        </x-table-cell>
+                    </tr>
                 </x-slot:foot>
             @endif
         </x-table>